% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{coef.PrevMap.ps}
\alias{coef.PrevMap.ps}
\title{Extract model coefficients from geostatistical linear model with preferentially sampled locations}
\usage{
\method{coef}{PrevMap.ps}(object, ...)
}
\arguments{
\item{object}{an object of class "PrevMap.ps".}

\item{...}{other arguments.}
}
\value{
a list of coefficients extracted from the model in \code{object}.
}
\description{
\code{coef} extracts parameters estimates from models fitted with the functions \code{\link{lm.ps.MCML}}.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}
}
