% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{coef.PrevMap}
\alias{coef.PrevMap}
\title{Extract model coefficients}
\usage{
\method{coef}{PrevMap}(object, ...)
}
\arguments{
\item{object}{an object of class "PrevMap".}

\item{...}{other arguments.}
}
\value{
coefficients extracted from the model object \code{object}.
}
\description{
\code{coef} extracts parameters estimates from models fitted with the functions \code{\link{linear.model.MLE}} and \code{\link{binomial.logistic.MCML}}.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
