% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{contour.pred.PrevMap}
\alias{contour.pred.PrevMap}
\title{Contour plot of a predicted surface}
\usage{
\method{contour}{pred.PrevMap}(x, type = NULL, summary = "predictions", ...)
}
\arguments{
\item{x}{an object of class "pred.PrevMap".}

\item{type}{a character indicating the type of prediction to display: 'prevalence', 'odds', 'logit' or 'probit'.}

\item{summary}{character indicating which summary to display: 'predictions','quantiles', 'standard.errors' or 'exceedance.prob'; default is 'predictions'. If \code{summary="exceedance.prob"}, the argument \code{type} is ignored.}

\item{...}{further arguments passed to \code{\link{contour}}.}
}
\description{
\code{plot.pred.PrevMap} displays contours of predictions obtained from \code{\link{spatial.pred.linear.MLE}}, \code{\link{spatial.pred.linear.Bayes}},\code{\link{spatial.pred.binomial.MCML}} and \code{\link{spatial.pred.binomial.Bayes}}.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
