% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galicia.boundary.R
\docType{data}
\name{galicia.boundary}
\alias{galicia.boundary}
\title{Boundary of Galicia}
\format{
A data frame with 42315 rows and 2 variables
}
\usage{
data(galicia.boundary)
}
\description{
This data-set contains the geographical coordinates of the boundary of the Galicia
region in northern Spain.

The variables are as follows:
\itemize{
 \item x x-coordinate of the spatial locations.
 \item y y-coordinate of the spatial locations.
}
}
\keyword{datasets}
