% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{loglik.ci}
\alias{loglik.ci}
\title{Profile likelihood confidence intervals}
\usage{
loglik.ci(object, coverage = 0.95, plot.spline.profile = TRUE)
}
\arguments{
\item{object}{object of class "profile.PrevMap" obtained from \code{\link{loglik.linear.model}}.}

\item{coverage}{a value between 0 and 1 indicating the coverage of the confidence interval based on the interpolated profile likelihood. Default is \code{coverage=0.95}.}

\item{plot.spline.profile}{logical; if \code{TRUE} an interpolating spline of the profile-likelihood of for a univariate parameter is plotted. Default is \code{FALSE}.}
}
\value{
A list with elements \code{lower} and \code{upper} for the upper and lower limits of the confidence interval, respectively.
}
\description{
Computes confidence intervals based on the interpolated profile likelihood computed for a single covariance parameter.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
