% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{plot.pred.PrevMap.ps}
\alias{plot.pred.PrevMap.ps}
\title{Plot of a predicted surface of geostatistical linear fits with preferentially sampled locations}
\usage{
\method{plot}{pred.PrevMap.ps}(x, target = NULL, summary = "predictions", ...)
}
\arguments{
\item{x}{an object of class "PrevMap".}

\item{target}{a integer value indicating the predictive target: \code{target=1} to visualize summaries of the surface associated with the response variable;
\code{target=2} to visualize summaries of the surface associated with the sampling intensity. If only one target has been predicted, this argument is ignored.}

\item{summary}{character indicating which summary to display: 'predictions','quantiles' or 'standard.errors'. Default is \code{summary='predictions'}. If \code{summary="exceedance.prob"}, the argument \code{type} is ignored.}

\item{...}{further arguments passed to \code{\link{plot}} of the 'raster' package.}
}
\description{
\code{plot.pred.PrevMap.ps} displays predictions obtained from \code{\link{lm.ps.MCML}}.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}
}
