% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{plot.PrevMap.diagnostic}
\alias{plot.PrevMap.diagnostic}
\title{Plot of the variogram-based diagnostics}
\usage{
\method{plot}{PrevMap.diagnostic}(x, ...)
}
\arguments{
\item{x}{an object of class "PrevMap.diagnostic".}

\item{...}{further arguments passed to \code{\link{plot}} of the 'raster' package.}
}
\description{
Displays the results from a call to \code{\link{variog.diagnostic.lm}} and \code{\link{variog.diagnostic.glgm}}.
}
\seealso{
\code{\link{variog.diagnostic.lm}}, \code{\link{variog.diagnostic.glgm}}
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
