% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{plot.profile.PrevMap}
\alias{plot.profile.PrevMap}
\title{Plot of the profile log-likelihood for the covariance parameters of the Matern function}
\usage{
\method{plot}{profile.PrevMap}(x, log.scale = FALSE, plot.spline.profile = FALSE, ...)
}
\arguments{
\item{x}{object of class "profile.PrevMap" obtained as output from \code{\link{loglik.linear.model}}.}

\item{log.scale}{logical; if \code{log.scale=TRUE}, the profile likleihood is plotted on the log-scale of the parameter values.}

\item{plot.spline.profile}{logical; if \code{TRUE} an interpolating spline of the profile-likelihood of for a univariate parameter is plotted. Default is \code{FALSE}.}

\item{...}{further arugments passed to \code{\link{plot}} if the profile log-likelihood is for only one parameter, or to \code{\link{contour}} for the bi-variate profile-likelihood.}
}
\value{
A plot is returned. No value is returned.
}
\description{
This function displays a plot of the profile log-likelihood that is computed by the function \code{\link{loglik.linear.model}}.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
