% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{plot.shape.matern}
\alias{plot.shape.matern}
\title{Plot of the profile likelihood for the shape parameter of the Matern covariance function}
\usage{
\method{plot}{shape.matern}(x, plot.spline = TRUE, ...)
}
\arguments{
\item{x}{an object of class 'shape.matern' obtained as result of a call to \code{\link{shape.matern}}}

\item{plot.spline}{logical; if \code{TRUE} an interpolating spline of the profile likelihood is added to the plot.}

\item{...}{further arguments passed to \code{\link{plot}}.}
}
\value{
The function does not return any value.
}
\description{
This function plots the profile likelihood for the shape parameter of the Matern covariance function using the output from \code{\link{shape.matern}} function.
}
\seealso{
\code{\link{shape.matern}}
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
