% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{point.map}
\alias{point.map}
\title{Point map}
\usage{
point.map(data, var.name, coords, ...)
}
\arguments{
\item{data}{an object of class "data.frame" containing the data.}

\item{var.name}{a \code{\link{formula}} object indicating the variable to display.}

\item{coords}{a \code{\link{formula}} object indicating the geographical coordinates.}

\item{...}{additional arguments to be passed to \code{points.geodata}.}
}
\description{
This function produces a plot with points indicating the data locations. Arguments can control the points sizes, patterns and colors. These can be set to be proportional to data values, ranks or quantiles. Alternatively, points can be added to the current plot.
}
