% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{trace.plot.MCML}
\alias{trace.plot.MCML}
\title{Trace-plots of the importance sampling distribution samples from the MCML method}
\usage{
trace.plot.MCML(object, component = NULL, ...)
}
\arguments{
\item{object}{an object of class "PrevMap" obatained as result of a call to \code{\link{binomial.logistic.MCML}}.}

\item{component}{a positive integer indicating the number of the random effect component for which a trace-plot is required. If \code{component=NULL}, then a component is selected at random. Default is \code{component=NULL}.}

\item{...}{further arguments passed to \code{\link{plot}}.}
}
\description{
Trace-plots of the MCMC samples from the importance sampling distribution used in \code{\link{binomial.logistic.MCML}}.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
