% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{trend.plot}
\alias{trend.plot}
\title{Plot of trends}
\usage{
trend.plot(data, var.name, coords, ...)
}
\arguments{
\item{data}{an object of class "data.frame" containing the data.}

\item{var.name}{a \code{\link{formula}} object indicating the variable to display.}

\item{coords}{a \code{\link{formula}} object indicating the geographical coordinates.}

\item{...}{additional arguments to be passed to \code{\link{plot}}.}
}
\description{
This function produces a plot of the variable of interest against each of the two geographical coordinates.
}
