% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loglik.ci}
\alias{loglik.ci}
\title{Profile log-likelihood confidence intervals}
\usage{
loglik.ci(object, coverage = 0.95, plot.spline.profile = TRUE)
}
\arguments{
\item{object}{object of class "profile.PrevMap" obtained from \code{\link{loglik.linear.model}}.}

\item{coverage}{confidence level; default is 0.95.}

\item{plot.spline.profile}{logical; if \code{TRUE} an interpolating spline of the profile-likelihood of for a univariate parameter is plotted. Default is \code{FALSE}.}
}
\value{
A list with elements \code{lower} and \code{upper} for the upper and lower limits of the confidence interval, respectively.
}
\description{
Computes confidence intervals based on the profile log-likelihood computed for a single covariance parameter.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}

