% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.pred.PrevMap}
\alias{plot.pred.PrevMap}
\title{Plot of a predicted surface}
\usage{
\method{plot}{pred.PrevMap}(x, type = NULL, summary = "predictions", ...)
}
\arguments{
\item{x}{an object of class "PrevMap".}

\item{type}{a character indicating the type of prediction to display: 'prevalence','odds' or 'logit'.}

\item{summary}{character indicating which summary to display: 'predictions','quantiles', 'standard.errors' or 'exceedance.prob'; default is 'predictions'. If \code{summary="exceedance.prob"}, the argument \code{type} is ignored.}

\item{...}{further arguments passed to \code{\link{plot}}.}
}
\description{
\code{plot.pred.PrevMap} displays predictions obtained from \code{\link{spatial.pred.linear.MLE}}, \code{\link{spatial.pred.linear.Bayes}},\code{\link{spatial.pred.binomial.MCML}} and \code{\link{spatial.pred.binomial.Bayes}}.
}
\author{
Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}

